add source code for WT_Perf and related archive
WT_Perf is the blade-element-momentum-theory code used by HARP_Opt.
WT_Perf is a Fortran code whose development dates back to 1985 and
possibly earlier.  This is the most recent version of WT_Perf v3.05.00a-adp
known since NREL discontinued support for this code in 2014.

Contained in this repository is the source code for WT_Perf, and
the NWTC_Subroutine_Library needed to compile the code. Makefiles,
user guide, and example test cases are included here.

long live WT_Perf!